requirements.txt
autopep8==1.5.4
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy==0.790
mypy-extensions==0.4.3
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
requests==2.25.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
